[GITHUB-60] Stops restarts during config restore
Removes notifies for GoCD restarts from configure.yml and makes
a Server restart only happen if a restore from a backup is not
taking place. This stops a potential issue where the config is
being restored and a restart happens part way through, this
could leave the server in a bad state.