[GITHUB-64] Stops restart during service start
Restarting the Server whilst it is starting can cause GoCD's
internal config Git repo to become corrupted, this change
ensures that the Server is stopped by default and then started
once configuration has taken place, with a wait after service
start to ensure that the server is in a good state.

Also adds a fix for Docker Hub rate limiting to travis.yml and
fixes an issue with tmp files in Ansible.